sspi-rs: move common functions to utils module
