+++ src/lib.rs
@@ -1,7 +1,10 @@
+//! The Rust implementation of the Remote Desktop Protocol.
+
 // FromPrimitive and ToPrimitive causes clippy error, so we disable it until
 // https://github.com/rust-num/num-derive/issues/20 is fixed
 #![cfg_attr(feature = "cargo-clippy", allow(clippy::useless_attribute))]
 
+mod utils;
 mod ber;
 mod crypto;
