+++ src/utils.rs
@@ -1,0 +1,17 @@
+use byteorder::{LittleEndian, ReadBytesExt};
+
+pub fn string_to_utf16(value: &str) -> Vec<u8> {
+    value
+        .encode_utf16()
+        .flat_map(|i| i.to_le_bytes().to_vec())
+        .collect::<Vec<u8>>()
+}
+
+pub fn bytes_to_utf16_string(mut value: &[u8]) -> String {
+    let mut value_u16 = vec![0x00; value.len() / 2];
+    value
+        .read_u16_into::<LittleEndian>(value_u16.as_mut())
+        .expect("read_u16_into cannot fail at this point");
+
+    String::from_utf16_lossy(value_u16.as_ref())
+}
